refactor notifications api and organize frontend pages
