+++ frontend/src/user-routes.tsx
@@ -25,8 +25,8 @@ const MCPActivation = lazy(() => import("./pages/MCPActivation.tsx"));
 const MCPStatus = lazy(() => import("./pages/MCPStatus.tsx"));
 const MacroCalculatorPage = lazy(() => import("./pages/MacroCalculatorPage.tsx"));
 const MessageTemplates = lazy(() => import("./pages/MessageTemplates.tsx"));
-const Notifications = lazy(() => import("./pages/Notifications.tsx"));
-const NotificationsDemo = lazy(() => import("./pages/NotificationsDemo.tsx"));
+const Notifications = lazy(() => import("./pages/notifications/Notifications.tsx"));
+const NotificationsDemo = lazy(() => import("./pages/notifications/NotificationsDemo.tsx"));
 const NutritionPlans = lazy(() => import("./pages/NutritionPlans.tsx"));
 const ProgramEditor = lazy(() => import("./pages/ProgramEditor.tsx"));
 const Progress = lazy(() => import("./pages/Progress.tsx"));
